// src/enums.ts
export enum Attributes {
  COMMAND_LINE = 'lightstep.command_line',
  COMPONENT_NAME = 'lightstep.component_name',
  HOSTNAME = 'lightstep.hostname',
  TRACER_PLATFORM = 'lightstep.tracer_platform',
  TRACER_PLATFORM_VERSION = 'lightstep.tracer_platform_version',
  TRACER_VERSION = 'lightstep.tracer_version',
}